# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: BANCO Asset Marketplace API — Cars, Real Estate, Industrial
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    ClerkAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [INVALID_DATA, NOT_FOUND, UNAUTHORIZED, INTERNAL_ERROR, FORBIDDEN, RATE_LIMITED, INVALID_TOKEN, CONFLICT]
        message:
          type: string
    Meta:
      type: object
      properties:
        cursor:
          type: string
        has_next:
          type: boolean
        total:
          type: integer
...
    FeedItem:
      type: object
      description: >
        Immutable v1 contract — exactly 11 fields.
        No renames, no removals allowed in v1. Breaking changes go to /v2/*.
      required:
        - id
        - media_preview
        - price_display
        - title
        - location
        - trust_signal
        - has_video
        - is_sponsored
      properties:
        id:
          type: string
          description: Listing identifier (use for detail navigation).
        media_preview:
          type: string
        price_display:
          type: string
        installment_badge:
          type: string
          nullable: true
        title:
          type: string
        location:
          type: string
        urgency_signal:
          type: string
          nullable: true
        trust_signal:
          type: string
        smart_badge:
          type: string
          nullable: true
        has_video:
          type: boolean
        is_sponsored:
          type: boolean
        coordinates:
          $ref: "#/components/schemas/Coordinates"
          nullable: true
        best_offer_badge:
          type: string
          nullable: true
        industrial_type:
          type: string
          nullable: true
          description: Industrial sub-type (factory, warehouse, land, production_line, raw_material, machine) used for client-side category grouping. Null for non-industrial listings.
        origin_type:
          type: string
          nullable: true
          description: '"imported" when the listing is imported (else null) — drives the "مستورد / Imported" card badge on industrial + car listings.'
        category:
          type: string
          enum: [car, real_estate, industrial]
          nullable: true
          description: Additive. The listing's main section — lets clients adapt per-section UI (e.g. the save glyph) without re-deriving it.
        is_active:
          type: boolean
          nullable: true
          description: True when the listing status is "active" (publicly visible and promotable). Lets owner-facing surfaces built on FeedItem (e.g. the profile grid) gate the Promote control without fetching ListingDetail. Null when status is unknown.
        whatsapp_enabled:
          type: boolean
          nullable: true
          description: True when the seller opted this listing in to WhatsApp contact. Opt-in only (default false). Lets surfaces decide whether to surface a WhatsApp action. Null when unknown.
        created_at:
          type: string
          nullable: true
          description: Listing creation timestamp (ISO 8601). Owner-facing grids (profile, "my listings") render a "Listed <date>" caption. Null only when the source row lacks it; public feeds always carry it.
        is_request:
          type: boolean
          nullable: true
          description: True when this is a buyer "request/wanted" post (looking to buy) rather than a sale listing. Surfaces render a "طلب / Wanted" badge and the requests filter keys off it. Null when unknown.
        is_bookable:
          type: boolean
          nullable: true
          description: >-
            Additive. True when this listing is a furnished/daily rental (hotel
            model) that can be reserved from its detail. Lets the feed, search
            card, and map pin surface a "قابل للحجز / Bookable" affordance without
            fetching the detail. Null/false for everything else.
    Coordinates:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
        lng:
          type: number
...
paths:
...
  /v1/feed:
    get:
      operationId: getFeed
      summary: Get adaptive scored feed
      tags: [feed]
      parameters:
        - in: query
          name: cursor
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: category
          schema:
            type: string
            enum: [car, real_estate, industrial]
        - in: query
          name: is_request
          description: >-
            Buyer request/wanted filter. Omit to include both sale listings and
            requests. true = only requests; false = only sale listings.
          schema:
            type: boolean
        - in: query
          name: industrial_type
          description: >-
            Industrial subtype filter (only meaningful with category=industrial).
            A single subtype or a comma-separated list (a whole group), e.g.
            "factory,warehouse,land". Valid subtypes: factory, warehouse, land,
            production_line, raw_material, machine.
          schema:
            type: string
        - in: query
          name: condition
          description: Filter by condition (cars). Real data via listing_attributes.condition or specs.
          schema:
            type: string
            enum: [new, used]
        - in: query
          name: payment_plan
          description: >-
            Payment engine filter. installment = any non-cash plan; bank =
            bank_finance; direct = seller_installment; islamic = sharia-compliant
            non-cash plan.
          schema:
            type: string
            enum: [installment, bank, direct, islamic]
        - in: query
          name: property_type
          description: Filter real-estate by property type (villa, apartment, penthouse, duplex, ...).
          schema:
            type: string
        - in: query
          name: finishing_type
          description: Filter real-estate by finishing level.
          schema:
            type: string
        - in: query
          name: compound
          description: Real-estate located inside a compound (specs.compound).
          schema:
            type: boolean
        - in: query
          name: furnished
          description: Furnished real-estate (specs.furnished).
          schema:
            type: boolean
        - in: query
          name: offer_type
          description: Real-estate offer type — sale (تمليك) or rent (إيجار), from specs.offer_type.
          schema:
            type: string
            enum: [sale, rent]
        - in: query
          name: rental_term
          description: >-
            Rental system within rent — the country's legal/duration regime
            (Egypt: furnished_daily from 1 day, new_law up to 5 years, old_law up
            to 59 years; Gulf: annual_contract). Free string; the catalog is
            client-side and grows per country.
          schema:
            type: string
        - in: query
          name: market_country
          description: >-
            ISO 3166-1 alpha-2 market country (EG, SA, …). Filters inventory by
            specs.market_country; listings without the key are treated as EG.
          schema:
            type: string
            pattern: "^[A-Za-z]{2}$"
        - in: query
          name: material
          description: >-
            Filter raw-material commodity listings by specs.material
            (steel, aluminum, copper, …). Free string; catalog is client-side.
          schema:
            type: string
            maxLength: 40
        - in: query
          name: fuel_type
          description: Filter cars by fuel type (listing_attributes.fuel_type or specs).
          schema:
            type: string
            enum: [petrol, diesel, hybrid, electric, natural_gas]
        - in: query
          name: transmission
          description: Filter cars by transmission (listing_attributes.transmission or specs).
          schema:
            type: string
            enum: [manual, automatic, cvt]
        - in: query
          name: brand
          description: >-
            Filter cars by brand, matched against the English listing title
            (titles are canonical "<Brand> <Model> <Year>").
          schema:
            type: string
        - in: query
          name: model
          description: Filter cars by model, matched against the English listing title.
          schema:
            type: string
        - in: query
          name: min_year
          description: Minimum car model year (numeric specs.year).
          schema:
            type: integer
        - in: query
          name: max_year
          description: Maximum car model year (numeric specs.year).
          schema:
            type: integer
        - in: query
          name: industry
          description: Filter industrial listings by industry (listing_attributes.industry).
          schema:
            type: string
            enum: [food, beverage, plastic, textile, pharmaceutical, chemical, engineering, other]
        - in: query
          name: origin_type
          description: Filter industrial listings by origin (listing_attributes.origin_type).
          schema:
            type: string
            enum: [local, imported]
        - in: query
          name: session_id
          schema:
            type: string
      responses:
        "200":
          description: Feed items with cursor pagination
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/FeedItem"
                  error:
                    $ref: "#/components/schemas/ApiError"
                    nullable: true
                  meta:
                    $ref: "#/components/schemas/Meta"
...